#628-repo-details-page repository details page header and banner styles added
